split types out
