+++ pkg/osc/types.go
@@ -1,0 +1,11 @@
+package osc
+
+type OSCArg struct {
+	Type  string
+	Value any
+}
+
+type OSCMessage struct {
+	Address string
+	Args    []OSCArg
+}
